qmpi: Split tool registration and initialization into two phases
Rather than having the tool register itself and register all of the
functions it wants to intercept at once, do this in two stages. First
the tool provides a name, then it gets a callback to register the rest
of the functions.

This allows us to create an environment variable (QMPI_TOOL_LIST) that
allows the user to specify which tools they want to use and in which
order. This value should be a colon-separated string.